refactor: modularize deduplication logic into dedup.py
