+++ src/run.py
@@ -11,6 +11,7 @@ from .github_issue import find_or_create_issue, create_comment, close_other_dail
 from .github_issue import list_comments
 from .slack import post_to_slack
 from .utils import debug_log, slugify_case_name
+from .dedup import apply_deduplication
 from .courtlistener import (
     search_recent_documents,
     build_complaint_documents_from_hits,
@@ -127,206 +128,10 @@ def main() -> None:
    
 
     # =========================================================
-    # Baseline 비교 로직
+    # Baseline 비교 로직 (Modularized)
     # =========================================================
     comments = list_comments(owner, repo, gh_token, issue_no)
-    first_run_today = len(comments) == 0
-
-    if not first_run_today:
-        # =====================================================
-        # 🔒 안정형 테이블 기반 비교 로직 (모든 이전 댓글 대상)
-        # =====================================================
-
-        def extract_section(md_text: str, section_title: str) -> str:
-            lines = md_text.split("\n")
-            start = None
-            end = None
-            for i, line in enumerate(lines):
-                if line.strip().startswith(section_title):
-                    start = i + 1
-                    continue
-                if start and line.startswith("## "):
-                    end = i
-                    break
-            if start is None:
-                return ""
-            if end is None:
-                end = len(lines)
-            return "\n".join(lines[start:end])
-
-        def parse_table(section_md: str):
-            lines = [l for l in section_md.split("\n") if l.strip().startswith("|")]
-            if len(lines) < 3:
-                return [], [], ()
-
-            header = lines[0]
-            separator = lines[1]
-            rows = lines[2:]
-
-            def split_row(row_text: str):
-                # 정규식 (?<!\\)\| 를 사용하여 역슬래시로 이스케이프되지 않은 파이프만 분할
-                return [c.strip() for c in re.split(r'(?<!\\)\|', row_text.strip())[1:-1]]
-
-            header_cols = split_row(header)
-
-            parsed_rows = []
-            for row in rows:
-                cols = split_row(row)
-                if len(cols) == len(header_cols):
-                    parsed_rows.append(cols)
-                else:
-                    debug_log(f"Table row column mismatch: expected {len(header_cols)}, got {len(cols)}. Row: {row[:100]}...")
-
-            return header_cols, parsed_rows, (header, separator)
-
-        def extract_article_url(cell: str):
-            m = re.search(r"\((https?://[^\)]+)\)", cell)
-            if m:
-                return m.group(1).split("&hl=")[0]
-            return None
-
-        # -------------------------
-        # Base Snapshot Key Set 생성 (모든 이전 댓글 대상)
-        # -------------------------
-        base_article_set = set()
-        base_docket_set = set()
-
-        for comment in comments:
-            body = comment.get("body") or ""
-            
-            # News 처리
-            news_section_base = extract_section(body, "## 📰 News")
-            h_news, r_news, _ = parse_table(news_section_base)
-            if "제목" in h_news:
-                idx = h_news.index("제목")
-                for r in r_news:
-                    url = extract_article_url(r[idx])
-                    if url:
-                        base_article_set.add(url)
-            
-            # Cases 처리
-            recap_section_base = extract_section(body, "## ⚖️ Cases")
-            h_cases, r_cases, _ = parse_table(recap_section_base)
-            if "도켓번호" in h_cases:
-                idx = h_cases.index("도켓번호")
-                for r in r_cases:
-                    base_docket_set.add(r[idx])
-
-        # -------------------------
-        # 현재 md 처리
-        # -------------------------
-        current_md = md
-
-        # 외부 기사 처리
-        news_section = extract_section(current_md, "## 📰 News")
-        headers, rows, table_meta = parse_table(news_section)
-
-        new_article_count = 0
-        total_article_count = len(rows)
-
-        if headers and "제목" in headers:
-            title_idx = headers.index("제목")
-            no_idx = headers.index("No.") if "No." in headers else None
-            date_idx = headers.index("기사일자⬇️") if "기사일자⬇️" in headers else None
-
-            header_line, separator_line = table_meta
-            
-            non_skip_rows = []
-            skip_rows = []
-
-            for r in rows:
-                url = extract_article_url(r[title_idx])
-                if url in base_article_set:
-                    # 개선: 핵심 식별 컬럼(No, 기사일자, 제목)은 유지
-                    new_row = []
-                    for i, col in enumerate(r):
-                        if i in (no_idx, date_idx, title_idx):
-                            new_row.append(col)
-                        else:
-                            new_row.append("skip")
-                    skip_rows.append(new_row)
-                else:
-                    non_skip_rows.append(r)
-                    new_article_count += 1
-            
-            # 합치기: 신규 먼저, 기존(skip) 나중
-            final_rows = non_skip_rows + skip_rows
-            new_lines = [header_line, separator_line]
-            
-            for row_idx, r in enumerate(final_rows, start=1):
-                if no_idx is not None:
-                    r[no_idx] = str(row_idx)
-                new_lines.append("| " + " | ".join(r) + " |")
-
-            new_news_section = "\n".join(new_lines)
-            current_md = current_md.replace(news_section, new_news_section)
-
-        # Cases 처리
-        recap_section = extract_section(current_md, "## ⚖️ Cases")
-        headers, rows, table_meta = parse_table(recap_section)
-
-        new_docket_count = 0
-        total_docket_count = len(rows)
-
-        if headers and "도켓번호" in headers:
-            docket_idx = headers.index("도켓번호")
-            no_idx = headers.index("No.") if "No." in headers else None
-            status_idx = headers.index("상태") if "상태" in headers else None
-            case_idx = headers.index("케이스명") if "케이스명" in headers else None
-
-            header_line, separator_line = table_meta
-            
-            non_skip_rows = []
-            skip_rows = []
-
-            for r in rows:
-                docket = r[docket_idx]
-                if docket in base_docket_set:
-                    # 개선: 핵심 식별 컬럼(No, 상태, 케이스명, 도켓번호) 유지
-                    new_row = []
-                    for i, col in enumerate(r):
-                        if i in (no_idx, status_idx, case_idx, docket_idx):
-                            new_row.append(col)
-                        else:
-                            new_row.append("skip")
-                    skip_rows.append(new_row)
-                else:
-                    non_skip_rows.append(r)
-                    new_docket_count += 1
-
-            # 합치기: 신규 먼저, 기존(skip) 나중
-            final_rows = non_skip_rows + skip_rows
-            new_lines = [header_line, separator_line]
-            
-            for row_idx, r in enumerate(final_rows, start=1):
-                if no_idx is not None:
-                    r[no_idx] = str(row_idx)
-                new_lines.append("| " + " | ".join(r) + " |")
-
-            new_recap_section = "\n".join(new_lines)
-            current_md = current_md.replace(recap_section, new_recap_section)
-
-        # -------------------------
-        # Summary 생성
-        # -------------------------
-        base_news = len(base_article_set)
-        base_cases = len(base_docket_set)
-
-        dup_news = total_article_count - new_article_count
-        dup_cases = total_docket_count - new_docket_count
-
-        summary_header = (
-            "### 중복 제거 요약:\n"
-            "🔁 Dedup Summary\n"
-            f"└ News {base_news} (Baseline): "
-            f"{dup_news} (Dup), "
-            f"{new_article_count} (New)\n"
-            f"└ Cases {base_cases} (Baseline): "
-            f"{dup_cases} (Dup), "
-            f"{new_docket_count} (New)\n\n"
-        )
-
-        md = summary_header + current_md
+    md = apply_deduplication(md, comments)
 
     # 이전 날짜 이슈 Close
     closed_nums = close_other_daily_issues(owner, repo, gh_token, issue_label, base_title, issue_title, issue_no, issue_url)
